+++ src/dotnet/ReSharperPlugin.SpecflowRiderPlugin/ProjectTemplateParameters/FluentAssertionsProviderParameter.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
-using JetBrains.ReSharper.Host.Features.ProjectModel.ProjectTemplates.DotNetExtensions;
-using JetBrains.ReSharper.Host.Features.ProjectModel.ProjectTemplates.DotNetTemplates;
+using JetBrains.Rider.Backend.Features.ProjectModel.ProjectTemplates.DotNetExtensions;
+using JetBrains.Rider.Backend.Features.ProjectModel.ProjectTemplates.DotNetTemplates;
 using JetBrains.Rider.Model;
 
 namespace ReSharperPlugin.SpecflowRiderPlugin.ProjectTemplateParameters