+++ src/dotnet/ReSharperPlugin.SpecflowRiderPlugin/ProjectTemplateParameters/TargetFrameworkProviderParameter.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
-using JetBrains.ReSharper.Host.Features.ProjectModel.ProjectTemplates.DotNetExtensions;
-using JetBrains.ReSharper.Host.Features.ProjectModel.ProjectTemplates.DotNetTemplates;
+using JetBrains.Rider.Backend.Features.ProjectModel.ProjectTemplates.DotNetExtensions;
+using JetBrains.Rider.Backend.Features.ProjectModel.ProjectTemplates.DotNetTemplates;
 using JetBrains.Rider.Model;
 
 namespace ReSharperPlugin.SpecflowRiderPlugin.ProjectTemplateParameters