+++ CHANGELOG.md
@@ -4,6 +4,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.9.4
+### Added
+- Support for Rider 2021.2 final
+
+## 1.9.3
+### Added
+- Support for Rider 2021.9 EAP build 212.4746.97
+- Update the project template to the latest SpecFlow version
+
+### Fixed
+- "Where to create the step?" UI isn't consistent [#89](https://github.com/SpecFlowOSS/SpecFlow.Rider/issues/89)
+
+## 1.9.2
+### Added
+- Support for Rider 2021.7 EAP build 212.4746.33
+- Support the new "Reformat and Clenaup" feature
+- Use the native notification popup for the on boarding messages
+
+### Fixed
+- Keyword completion is case sensitive by default [#148](https://github.com/SpecFlowOSS/SpecFlow.Rider/issues/148)
+
+## 1.9.1
+### Added
+- Support for Rider 2021.2 EAP build 212.4416.21
+
+## 1.9.0
+### Added
+- Support for Rider 2021.2 EAP build 212.4037.61
+
 ## 1.8.0
 
 ### Added